backend/requirements2.txt
airportsdata==20250909
alembic==1.18.4
APScheduler==3.11.0
asyncpg==0.31.0
authx==1.5.1
Babel==2.18.0
bcrypt==5.0.0
email-validator==2.3.0
fastapi==0.135.1
google_search_results==2.4.2
langchain_google_genai==4.2.1
langgraph==1.0.10
psycopg_pool==3.3.0
pydantic==2.12.5
pydantic_settings==2.13.1
PyJWT==2.11.0
python-dotenv==1.2.2
redis==7.3.0
Requests==2.32.5
SQLAlchemy==2.0.48
typing_extensions==4.15.0
uvicorn==0.41.0
